+++ libs/mock-example/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.6.0-rc.0](https://github.com/dynamicmsw/dynamic-msw/compare/v1.5.1...v1.6.0-rc.0) (2022-12-20)
+
+
+### Features
+
+* **core:** allow for non object select options ([#117](https://github.com/dynamicmsw/dynamic-msw/issues/117)) ([a97ec2a](https://github.com/dynamicmsw/dynamic-msw/commit/a97ec2a986f6aac9cf49e74b0707eb673ff03d8e))
+
 ### [1.5.1](https://github.com/dynamicmsw/dynamic-msw/compare/v1.5.0...v1.5.1) (2022-11-21)
 
 ## [1.5.0](https://github.com/dynamicmsw/dynamic-msw/compare/v1.4.0...v1.5.0) (2022-11-20)